ai-model/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-generativeai
numpy
opencv-python-headless